Robo Jumper V1
The is the first/starting version of my 3D platformer, the main character can run, collect coins, and the bonus point system works like a dream. I also have a damage cube set up at the moment, this is purely just to test the character taking damage and dying. I also have the main menu, the main UI and the ability to pause the game and resume working. One more this this version currently has an unused folder label GTVersion. This is only here to help distinguish this version from the original version on my computer, will likely be removed in the future