Robo Jumper V3
For this version I reworked the jumping/is grounded checking, so that the jump is more responsive and the game does a better job tracking weather or not the player is currently on the ground. Plus they also now go into a falling animation when falling, I also made it so that so once a level is completed it grades the player with a total score based on the overall score, Plus bonus points depending on how many lives they still have, and the number of coins they collected. I ended creating the images of the medals with Figma. I also rearranged the point, coin and lives variables to be an a separate script so that they can tracked and read across all scenes/menus, and made it so that all instances of these variables being used all called from that script.