Robo Jumper V4
For this version I did a few things, firstly I added an animation for falling into the pit (Which will basically trigger if the player dies in mid air), along with adding in an animation when landing, which then transitions back into either standing, walking or running. I also created an enemy  with simple AI to navigate a set area and chase/shoot the player if they detect them, along with the player taking damage if they're shot. There are 2 enemies, one on a separate platform with 3 pillars, and 1 that navigates a boxed of area.