RoboJumper V5 (Note, Final prototye version)
I added audio to the game, right now there are 2 sound files, one for collecting coins and reaching the end level goal (Currently reusing the same sound for both), and a sound for taking damage. I also added in the options for both the main menu and the in game pause menu. In the main menu the user can mute/unmute the games audio and disable weather or not ranks are shown at the end of a level (Still shows the total score , just doesn't give them a medal). As of this version there arn't any other options in pause menu)